+++ examples/webpack/client/main.js
@@ -1,0 +1,6 @@
+var div = document.querySelector("#counter");
+
+var counter = 1;
+setInterval(function() {
+	div.textContent = "Counter " + counter++;
+}, 1000);